Move NumericSupport to macro-utils project where CaseClass1Rep is
