+++ macro-utils/src/main/scala-2.12/pl/iterators/kebs/support/NumericSupport.scala
@@ -1,4 +1,4 @@
-package pl.iterators.kebs.tagged.implicits
+package pl.iterators.kebs.support
 
 import pl.iterators.kebs.macros.CaseClass1Rep
