+++ macro-utils/src/main/scala/pl/iterators/kebs/support/package.scala
@@ -1,0 +1,3 @@
+package pl.iterators.kebs
+
+package object support extends NumericSupport
