--- tagged/src/main/scala/pl/iterators/kebs/tagged/implicits/package.scala
@@ -1,3 +1,0 @@
-package pl.iterators.kebs.tagged
-
-package object implicits extends NumericSupport
